+++ R/explore.R
@@ -217,6 +217,7 @@ explore <- function(
   speed.warning <- aux$speed.warning
   speed.error <- aux$speed.error
   jump.warning <- aux$jump.warning
+  jump.error <- aux$jump.error
   inactive.warning <- aux$inactive.warning
   inactive.error <- aux$inactive.error
   detections.y.axis <- aux$detections.y.axis