+++ R/migration.R
@@ -207,6 +207,7 @@ migration <- function(
   speed.warning <- aux$speed.warning
   speed.error <- aux$speed.error
   jump.warning <- aux$jump.warning
+  jump.error <- aux$jump.error
   inactive.warning <- aux$inactive.warning
   inactive.error <- aux$inactive.error
   detections.y.axis <- aux$detections.y.axis