+++ R/residency.R
@@ -204,6 +204,7 @@ residency <- function(
   speed.warning <- aux$speed.warning
   speed.error <- aux$speed.error
   jump.warning <- aux$jump.warning
+  jump.error <- aux$jump.error
   inactive.warning <- aux$inactive.warning
   inactive.error <- aux$inactive.error
   detections.y.axis <- aux$detections.y.axis